+++ src/entities/SearchItem.ts
