+++ src/frontend/containers/AdvancedSearch/SearchItemDialog.tsx
@@ -1,7 +1,7 @@
 import { observer } from 'mobx-react-lite';
 import React, { useState, useRef, useCallback } from 'react';
 import { ID } from 'src/api/ID';
-import { ClientFileSearchItem } from 'src/entities/FileSearchItem';
+import { ClientFileSearchItem } from 'src/entities/SearchItem';
 import { useStore } from 'src/frontend/contexts/StoreContext';
 import { useAutorun } from 'src/frontend/hooks/mobx';
 import { Button } from 'widgets/Button';
